Split swarm command panels
